Migrate camera APIs to separate tag
