+++ frigate/api/fastapi_app.py
@@ -15,6 +15,7 @@ from starlette_context.plugins import Plugin
 from frigate.api import app as main_app
 from frigate.api import (
     auth,
+    camera,
     classification,
     event,
     export,
@@ -114,6 +115,7 @@ def create_fastapi_app(
     # Routes
     # Order of include_router matters: https://fastapi.tiangolo.com/tutorial/path-params/#order-matters
     app.include_router(auth.router)
+    app.include_router(camera.router)
     app.include_router(classification.router)
     app.include_router(review.router)
     app.include_router(main_app.router)
